GitHub test and release actions (#270)
* add dependabot config

this will auto update github actions when new versions are available

Signed-off-by: Jeroen Simonetti <jeroen@simonetti.nl>

* Publish or test containers

Replace github workflow for building generic and addon container images.

Signed-off-by: Jeroen Simonetti <jeroen@simonetti.nl>

* Change addon config

Set default version to develop, as this will be automatically changed
by the github action.
Set the image to contain the -addon postfix.

Signed-off-by: Jeroen Simonetti <jeroen@simonetti.nl>

* Change the job name

The job name should reflect that it is also used for testing

Signed-off-by: Jeroen Simonetti <jeroen@simonetti.nl>

---------

Signed-off-by: Jeroen Simonetti <jeroen@simonetti.nl>