Initial version of config editor (#443)
* Initial version of config editor

Add a config ui editor under the /config route. For now I have chosen
to not expose that route, so it can be tested thouroughly before
exposing it to the public.

Features:
- changes are only saved when hitting the save button
- uses modal dialogs where applicable
- adds context sensitive help to each input field
- live-search for homeassistant entities (requires the homeassistant api
  configuration to be completed)

Added 5 new routes:
- /config: the guid config editor
- /api/settings/options.json: get and save options.json
- /api/settings/secrets.json: get and save secrets.json
- /api/ha/entities and /api/ha/entities/search: used for realtime entity
  search

Signed-off-by: Jeroen Simonetti <jeroen@simonetti.nl>

* Add max-height and overflow styles to battery tab, enables vertical scrolling

Add max-height and overflow styles to battery tab, enables vertical scrolling

* Refactor battery stage item layout with flexbox

Refactor stage layout to use flexbox for better alignment and spacing. Update HTML structure to group header and fields in a single row.

Compactere layout

* "Home assistant entity" ipv "Entity"

* Fix database server and username

The fields where mapped to the wrong json fields. This commit uses the
correct fields as specified in DOCS.md

Signed-off-by: Jeroen Simonetti <jeroen@simonetti.nl>

* Move stuff around

Also fix the dropdown for entso-e key to reference a secret.

Signed-off-by: Jeroen Simonetti <jeroen@simonetti.nl>

* Make charge/discharge stages more compact

This will take up less vertical screen realestate for the battery charge
and discharge stages and the ev discharge stages.

Signed-off-by: Jeroen Simonetti <jeroen@simonetti.nl>

* Make solar string scrollable

To prevent field from getting outside a screen the strings container is
now scrollable both horizontally and vertically

Signed-off-by: Jeroen Simonetti <jeroen@simonetti.nl>

* Add context help to all fields.

Some fields did not have a context help. This commit adds context (?)
help to every field.

Signed-off-by: Jeroen Simonetti <jeroen@simonetti.nl>

* Refactor the context help.

We can now add much more text in the help. Also allows for (simple) text
formatting.

Use \n for line breaks
Use **text** for bold text
Use _text_ for italic text
Use • for bullet points in the summary section

Look at 'general.strategy' for an example.

Signed-off-by: Jeroen Simonetti <jeroen@simonetti.nl>

* Add more context help

This is an AI translation of DOCS.md, some translation issues may be
present.

Signed-off-by: Jeroen Simonetti <jeroen@simonetti.nl>

* Fix help popup position

Sometimes the popup would go outside the viewport. This commit makes
sure the popup is always inside the viewport.

Signed-off-by: Jeroen Simonetti <jeroen@simonetti.nl>

* Add download and upload buttons

This commit adds buttons to download/upload options.json.

Signed-off-by: Jeroen Simonetti <jeroen@simonetti.nl>

* Add download and upload buttons

This commit adds buttons to download/upload options.json.

Signed-off-by: Jeroen Simonetti <jeroen@simonetti.nl>

* Add codeblock formating to context help.

Since it is often needed to refer to code or otherwise structed
information (such as a homeassistant entity), we add support for
backtics to indicate a code block.

Signed-off-by: Jeroen Simonetti <jeroen@simonetti.nl>

* Fix dutch wording leftovers

Signed-off-by: Jeroen Simonetti <jeroen@simonetti.nl>

---------

Signed-off-by: Jeroen Simonetti <jeroen@simonetti.nl>
Co-authored-by: balk77 <s.balkenende@gmail.com>